feat: Refactor para generar token y consultar carga/espacios
