+++ swagger/swagger.json
@@ -13,27 +13,58 @@
                 "tags": [
                     "carga_academica"
                 ],
-                "description": "create CargaAcademica",
+                "description": "query CargaAcademica",
                 "operationId": "CargaAcademicaController.PostCargaAcademica",
                 "parameters": [
                     {
                         "in": "body",
                         "name": "body",
-                        "description": "body for CargaAcademica content",
+                        "description": "body for Parametros content",
                         "required": true,
                         "schema": {
-                            "$ref": "#/definitions/models.CargaAcademica"
+                            "$ref": "#/definitions/models.Parametros"
                         }
                     }
                 ],
                 "responses": {
-                    "201": {
+                    "200": {
                         "description": "",
                         "schema": {
-                            "$ref": "#/definitions/models.CargaAcademica"
+                            "$ref": "#/definitions/models.Parametros"
                         }
                     },
-                    "403": {
+                    "404": {
+                        "description": "body is empty"
+                    }
+                }
+            }
+        },
+        "/espacios_academicos/": {
+            "post": {
+                "tags": [
+                    "espacios_academicos"
+                ],
+                "description": "query EspacioAcademico",
+                "operationId": "EspacioAcademicoController.PostEspacioAcademico",
+                "parameters": [
+                    {
+                        "in": "body",
+                        "name": "body",
+                        "description": "body for Parametros content",
+                        "required": true,
+                        "schema": {
+                            "$ref": "#/definitions/models.Parametros"
+                        }
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "",
+                        "schema": {
+                            "$ref": "#/definitions/models.Parametros"
+                        }
+                    },
+                    "404": {
                         "description": "body is empty"
                     }
                 }
@@ -208,10 +239,6 @@
         }
     },
     "definitions": {
-        "models.CargaAcademica": {
-            "title": "CargaAcademica",
-            "type": "object"
-        },
         "models.MetricasAutoevaluacion": {
             "title": "MetricasAutoevaluacion",
             "type": "object"
@@ -222,6 +249,10 @@
         },
         "models.MetricasHeteroevaluacion": {
             "title": "MetricasHeteroevaluacion",
+            "type": "object"
+        },
+        "models.Parametros": {
+            "title": "Parametros",
             "type": "object"
         },
         "models.Respuesta_formulario": {
@@ -245,6 +276,10 @@
         {
             "name": "carga_academica",
             "description": "CargaAcademicaController operations for CargaAcademica\n"
+        },
+        {
+            "name": "espacios_academicos",
+            "description": "EspacioAcademicoController operations for EspacioAcademico\n"
         }
     ]
 }
