+++ swagger/swagger.yml
@@ -9,21 +9,41 @@ paths:
     post:
       tags:
       - carga_academica
-      description: create CargaAcademica
+      description: query CargaAcademica
       operationId: CargaAcademicaController.PostCargaAcademica
       parameters:
       - in: body
         name: body
-        description: body for CargaAcademica content
+        description: body for Parametros content
         required: true
         schema:
-          $ref: '#/definitions/models.CargaAcademica'
+          $ref: '#/definitions/models.Parametros'
       responses:
-        "201":
+        "200":
           description: ""
           schema:
-            $ref: '#/definitions/models.CargaAcademica'
-        "403":
+            $ref: '#/definitions/models.Parametros'
+        "404":
+          description: body is empty
+  /espacios_academicos/:
+    post:
+      tags:
+      - espacios_academicos
+      description: query EspacioAcademico
+      operationId: EspacioAcademicoController.PostEspacioAcademico
+      parameters:
+      - in: body
+        name: body
+        description: body for Parametros content
+        required: true
+        schema:
+          $ref: '#/definitions/models.Parametros'
+      responses:
+        "200":
+          description: ""
+          schema:
+            $ref: '#/definitions/models.Parametros'
+        "404":
           description: body is empty
   /formulario_por_tipo/:
     get:
@@ -134,9 +154,6 @@ paths:
         "403":
           description: body is empty
 definitions:
-  models.CargaAcademica:
-    title: CargaAcademica
-    type: object
   models.MetricasAutoevaluacion:
     title: MetricasAutoevaluacion
     type: object
@@ -145,6 +162,9 @@ definitions:
     type: object
   models.MetricasHeteroevaluacion:
     title: MetricasHeteroevaluacion
+    type: object
+  models.Parametros:
+    title: Parametros
     type: object
   models.Respuesta_formulario:
     title: Respuesta_formulario
@@ -162,3 +182,6 @@ tags:
 - name: carga_academica
   description: |
     CargaAcademicaController operations for CargaAcademica
+- name: espacios_academicos
+  description: |
+    EspacioAcademicoController operations for EspacioAcademico
